Initial commmit of the www-site-checker project.
This project currently provides a (WIP) CLI tool which checks all of the
outbound links featured in the configured website(s), comparing them with
a pre-configured allowlist.

At this point, the tool runs from the CLI on a local machine. The goal is to
run it from CI, so subsequent commits will focus on getting that running,
followed by further tuning of the code and the allowlist, to make it all run
with a sufficient level of performance.

Tests: to come
Review: to come