Initial work: create site checking tool and hook into a GH Action (#1)
* Initial commmit of the www-site-checker project.

This project currently provides a (WIP) CLI tool which checks all of the
outbound links featured in the configured website(s), comparing them with
a pre-configured allowlist.

At this point, the tool runs from the CLI on a local machine. The goal is to
run it from CI, so subsequent commits will focus on getting that running,
followed by further tuning of the code and the allowlist, to make it all run
with a sufficient level of performance.

Tests: to come
Review: to come

* Switching output files to a local dir

* Update GH Action to upload artefacts; add retry support

* Add yaml check to pre-commit hooks

* Log to Sentry if unexpected URLs are found

This is rather than throwing an exception because that stops the GH Action from
copying the files to the artefacts store :facepalm:

* Fixup output filename to not clash with Github Actions rule

The colon in the timestamp from isoformat() is forbidden in GHA

* Dial up Sentry message to Warning level, rather than info

* Support batching the total list of pages to check

...so we can parallelize this in GH Actions

* Trying out batching in GH Action, initially starting with four batches

* Update allowlist to trust mozilla.com too

* Dial up Sentry logging to try to make message more reliably sent to subscribers

* Support passing a path to a custom allowlist via CLI args

* Drop '--nodump' option -- not useful

* Update US Origin server domain - was incorrect

* Update URL-harvesting code to update the hostname in a better place

Before, hostname replacement was happening too late in the process, meaning
a nested sitemap would result in the default/CDN being used to get the
sub-sitemap. Moving the URL-rewriting logic to be called earlier helps.

Tests still needed; simplification desirable

* Improve _dump_to_file: only use when needed; include hostname in filename

* Add config to check origin servers, too

TODO: refactor this config file to DRY it up

* Simple Slack integration - bit chatty/spammy, but tolerable for short term

* Move Slack ping to a separate script, so we alert via Slack just once

* Update crontab for scheduled checks (but keep it disabled for now)

* Temporarily disable the Slack integration

* Update README now that most behaviour is implemented

* Add HTTPError to the exceptions we allow retries for

This _might_ be a bit too liberal, so we may walk it back (or add extra
checks like allowing retries only for certain HTTP status codes)

* Re-enable Slack integration

* Re-visit code, updating comments and docstrings

* Update allowlist to support checking localhost + update docs to explain --allow-hostname option

* Rephrase International English 'artefact' to be 'artifact', for consistency with how it's phrased in Github Actions

* Typo fixes and add missing local-install step

* Regex update: replace cumbersome http[s]{0,1}: with https?:

Thanks for the steer @robhudson!

* GH Action update: refactor to matrix strategy; call sys.exit if unexpected URLs found

* Update allowlist for firefox.com, mozilla.org and mozilla.com with no trailing slash

* Refine mozilla.com|org and firefox.com regexes to support links with no path or trailing slash

* Add Mozilla-specific mailto: and app download links to allowlist

* Broaden allowlist based on existing outbound links - sense-checking them as we go

* Add regexes for first-party/Mozilla domains

* Expand mobile-download regex and add non-regex link, too

* Add pre-canned Tweets and mailto: links to the allowlist

* Bulk addition of URLs to the allowlist based on URLs which are already in use

* Add some malformatted or broken URLs currently discovered in pages

We'll fix these up, but for now don't want them triggering 'false alarms' as
they are already in the site, so not 'unexpected'

* Add ConnectionError to the list of exceptions we retry for

* Further allowlist tuning, incl broadening mozillians.org regex

* Add a few more broken links to the allowlist, until we clean them up

* Add more adjust.com rules: this time for FF Klar

* Allow for linebreaks in some selected URls - temporary support only, though

* Temporarily disable allowlist exceptions, to get a report to help fix them

* Revert "Temporarily disable allowlist exceptions, to get a report to help fix them"

This reverts commit 8057b5836504cf5aa49fb9ab704bfd4e9f8e1752.

* Remove a couple of exceptions on the allowlist, because the source has been fixed upstream

* Update allowlist for exceptions/buggy links - almost all are WIP for a fix, just awaiting a PR

* Remove fixed broken URLs from exceptions on allowlist

* Further allowlist updates

Reinstates now-allowed, fixed outbound URLs from the HOF

The `url` allowance has to remain until translations are updated

* Enable scheduled checks

* Allow relative path (parent 'directory') for links

* Update allowlist

* Plumb in tests, using pytest + coverage via Makefile

* Permit running the tool against a sitemap not yet referenced in the allowlist

This avoids the need to add a hostname to the allowlist before checking a
site, making initial exploration of a sitemap easier. However, it will mean
that every single URL encountered is flagged

Resolves #3

```
$ ./bin/run_checks.py --sitemap-url=https://subdomain.example.com/blog/sitemap.xml
Seeking an appropriate allowlist in file data/allowlist.yaml
Could not find an allowlist for addons.mozilla.org, so treating all outbound URLs as unexpected
Adding 20 URLs
Discovered 20 URLs to check
Pulling down https://subdomain.example.com/blog/
...
[etc]
...
Unexpected outbound URLs found on subdomain.example.com!
...
```

* GH Action: always exit with a non-zero code if unexpected links found

* Update allowlist