Expand post-check actions to open issues or PRs depending on the URLs discovered
If a URL can just be added to the allowlist, we raise a PR that adds it/them
to the allowlist and raise a PR via the github cli, which is available in a
GH Action context.

If a URL is malformed and so can't be safely added to the list, we open an
issue to investigate it.