Ensure requests.get() doesn't follow redirects
It's an edge case, but if our sitemap contains links that then 30x to
another site, we'll still scan the HTML of those pages (but we don't
spider any deeper than that).

We don't want to follow redirects from pages in our sitemap, as we end up
with unexpected results, and could end up adding URLs to the allowlist
that we don't need.

This change stops the following of redirects, and the response.text
involved with 30x responses is an empty string, which is fine for all
the processing which is attempted on responses, so we don't need to add
complexity of status-code checking right now.